Pokemon-_API/requirements.txt
Flask 
datetime
Flask-SQLAlchemy
PyJWT